+++ plugins/main/public/components/common/data-source/pattern/system-inventory/data-source-repository.ts
@@ -1,0 +1,49 @@
+import { createPatternDataSourceRepositoryUseSettingValue } from '../pattern-data-source-repository-use-setting-value';
+
+export const SystemInventoryStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue('system_inventory.pattern');
+
+export const SystemInventoryNetworksStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_networks.pattern',
+  );
+
+export const SystemInventoryInterfacesStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_interfaces.pattern',
+  );
+
+export const SystemInventoryProtocolsStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_protocols.pattern',
+  );
+
+export const SystemInventoryProcessesStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_processes.pattern',
+  );
+
+export const SystemInventoryPortsStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_ports.pattern',
+  );
+
+export const SystemInventoryPackagesStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_packages.pattern',
+  );
+
+export const SystemInventoryHotfixesStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_hotfixes.pattern',
+  );
+
+export const SystemInventorySystemStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_system.pattern',
+  );
+
+export const SystemInventoryHardwareStatesDataSourceRepository =
+  createPatternDataSourceRepositoryUseSettingValue(
+    'system_inventory_hardware.pattern',
+  );