+++ plugins/main/public/components/overview/it-hygiene/packages/inventories/hotfixes/managed-filters.ts
@@ -1,0 +1,7 @@
+export default [
+  {
+    type: 'multiSelect',
+    key: 'package.hotfix.name',
+    placeholder: 'Hotfix',
+  },
+];