+++ plugins/main/public/components/overview/it-hygiene/processes/inventories/processes/managed-filters.ts
@@ -1,0 +1,17 @@
+export default [
+  {
+    type: 'multiSelect',
+    key: 'process.name',
+    placeholder: 'Name',
+  },
+  {
+    type: 'multiSelect',
+    key: 'process.state',
+    placeholder: 'State',
+  },
+  {
+    type: 'multiSelect',
+    key: 'process.command_line',
+    placeholder: 'Command line',
+  },
+];