+++ plugins/main/public/components/overview/vulnerabilities/dashboards/inventory/inventory.tsx
@@ -34,7 +34,11 @@ import { LoadingSearchbarProgress } from '../../../../../../public/components/co
 // common components/hooks
 import useSearchBar from '../../../../common/search-bar/use-search-bar';
 import { useDataGrid } from '../../../../common/data-grid/use-data-grid';
-import { withErrorBoundary } from '../../../../common/hocs';
+import {
+  HideOnErrorInitializatingDataSource,
+  PromptErrorInitializatingDataSource,
+  withErrorBoundary,
+} from '../../../../common/hocs';
 import { exportSearchToCSV } from '../../../../common/data-grid/data-grid-service';
 import { compose } from 'redux';
 import { withVulnerabilitiesStateDataSource } from '../../common/hocs/validate-vulnerabilities-states-index-pattern';
@@ -56,6 +60,8 @@ import VulsEvaluationFilter, {
   getUnderEvaluationFilterValue,
 } from '../../common/components/vuls-evaluation-filter';
 import { DataGridVisibleColumnsSelector } from '../../../../common/wazuh-discover/components/visible-columns-selector';
+import { SampleDataWarning } from '../../../../visualize/components';
+import { WAZUH_SAMPLE_VULNERABILITIES } from '../../../../../../common/constants';
 
 const InventoryVulsComponent = () => {
   const {
@@ -66,6 +72,7 @@ const InventoryVulsComponent = () => {
     isLoading: isDataSourceLoading,
     fetchData,
     setFilters,
+    error,
   } = useDataSource<tParsedIndexPattern, PatternDataSource>({
     DataSource: VulnerabilitiesDataSource,
     repository: new VulnerabilitiesDataSourceRepository(),
@@ -115,6 +122,7 @@ const InventoryVulsComponent = () => {
   ]);
 
   const dataGridProps = useDataGrid({
+    moduleId: 'vulnerabilities-inventory',
     ariaLabelledBy: 'Vulnerabilities Inventory Table',
     defaultColumns: inventoryTableDefaultColumns,
     renderColumns: wzDiscoverRenderColumns,
@@ -171,6 +179,11 @@ const InventoryVulsComponent = () => {
         ErrorHandler.handleError(searchError);
       });
   }, [
+    /* FIX: If changing any filter and the results total is the same, the page is not reset to the
+     initial page but it causes a request. This has a different behavior compared with the table
+     used for Events. We should use the common logic for all the tables instead of creating new ones
+     avoinding the differences in the stuff that should be common for all the tables.
+     See the generic table plugins/main/public/components/common/wazuh-discover/table.tsx */
     JSON.stringify(fetchFilters),
     JSON.stringify(query),
     JSON.stringify(pagination),
@@ -210,6 +223,9 @@ const InventoryVulsComponent = () => {
     <IntlProvider locale='en'>
       <>
         <ModuleEnabledCheck />
+        {/* TODO: Using a page template wrapping these components causes different y render position
+        of data source error prompt. We should unify the different views. In the Dashboard tab, the
+        same prompt is rendered at top of view */}
         <EuiPageTemplate
           className='vulsInventoryContainer'
           restrictWidth='100%'
@@ -222,24 +238,31 @@ const InventoryVulsComponent = () => {
             {isDataSourceLoading ? (
               <LoadingSearchbarProgress />
             ) : (
-              <WzSearchBar
-                appName='inventory-vuls'
-                {...searchBarProps}
-                filters={excludeUnderEvaluationFilter(filters)}
-                fixedFilters={fixedFilters}
-                postFixedFilters={[
-                  () => (
-                    <VulsEvaluationFilter
-                      value={underEvaluation}
-                      setValue={handleFilterChange}
-                    />
-                  ),
-                ]}
-                showDatePicker={false}
-                showQueryInput={true}
-                showQueryBar={true}
-                showSaveQuery={true}
-              />
+              <>
+                <HideOnErrorInitializatingDataSource error={error}>
+                  <WzSearchBar
+                    appName='inventory-vuls'
+                    {...searchBarProps}
+                    filters={excludeUnderEvaluationFilter(filters)}
+                    fixedFilters={fixedFilters}
+                    postFixedFilters={[
+                      () => (
+                        <VulsEvaluationFilter
+                          value={underEvaluation}
+                          setValue={handleFilterChange}
+                        />
+                      ),
+                    ]}
+                    showDatePicker={false}
+                    showQueryInput={true}
+                    showQueryBar={true}
+                    showSaveQuery={true}
+                  />
+                  <SampleDataWarning
+                    categoriesSampleData={[WAZUH_SAMPLE_VULNERABILITIES]}
+                  />
+                </HideOnErrorInitializatingDataSource>
+              </>
             )}
             {!isDataSourceLoading && results?.hits?.total === 0 ? (
               <DiscoverNoResults />
@@ -329,6 +352,7 @@ const InventoryVulsComponent = () => {
               </EuiFlyout>
             )}
           </>
+          {error && <PromptErrorInitializatingDataSource error={error} />}
         </EuiPageTemplate>
       </>
     </IntlProvider>